backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-adk
google-genai
pypdf
